% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{keycloak_shiny_login_server}
\alias{keycloak_shiny_login_server}
\title{Módulo de Shiny para iniciar sesión con API de AIS}
\usage{
keycloak_shiny_login_server(config, redirect_uri, auto_redirect = TRUE)
}
\description{
Este módulo de Shiny hace varias cosas:
\enumerate{
\item Si no hay token, inicia el proceso de autorización.
\item Crea un objeto de conexión a la API de AIS y la almacenará en \code{session$userData$conn}.
\item Devuelve un \code{reactive} con un promise del perfil del usuario.
}

Para utilizarlo se debe llamar esta función al inicio de la función
\code{server} de la aplicación de Shiny junto a sus argumentos.
}
